# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Open Api",
        "contact": {
            "email": "admin@juzaweb.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/post-type/{type}": {
            "get": {
                "tags": [
                    "Post Type"
                ],
                "summary": "Get list post type items",
                "operationId": "v1.post-type.type.index",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/query_limit"
                    },
                    {
                        "$ref": "#/components/parameters/query_keyword"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success_list"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            }
        },
        "/api/post-type/{type}/{id}": {
            "get": {
                "tags": [
                    "Post Type"
                ],
                "summary": "Get post type data by id",
                "operationId": "v1.post-type.type.show",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/id_in_path"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success_detail"
                    },
                    "404": {
                        "$ref": "#/components/responses/error_404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error_500"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "success_detail": {
                "description": "Get Data Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "success_list": {
                "description": "Get List Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "error_404": {
                "description": "Page not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "error_500": {
                "description": "Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id_in_path": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "query_limit": {
                "name": "limit",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
...
            "query_keyword": {
                "name": "keyword",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer :token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}